requirements.txt
pip
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.19.0
aiozipstream==0.4
altgraph==0.17.3
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
auto-py-to-exe==2.37.0
beautifulsoup4==4.12.2
bottle==0.12.25
bottle-websocket==0.2.9
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
choicelib==0.1.5
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
distlib==0.3.7
Eel==0.16.0
filelock==3.12.4
fonttools==4.42.0
frozenlist==1.4.0
future==0.18.3
gevent==23.7.0
gevent-websocket==0.10.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
kiwisolver==1.4.4
line-profiler==4.1.0
loguru==0.7.0
magic-filter==1.0.11
markovify==0.9.4
matplotlib==3.7.2
multidict==6.0.4
numpy==1.25.2
packaging==23.1
pefile==2023.2.7
pickleDB==0.9.2
Pillow==10.0.0
platformdirs==3.10.0
pprofile==2.1.0
psutil==5.9.5
py-heat==0.0.6
pyaes==1.6.1
pyasn1==0.5.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
PyMarkovChain==1.8
pyparsing==3.0.9
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-vlc==3.0.18122
pywin32-ctypes==0.2.2
requests==2.31.0
rsa==4.9
schedule==1.2.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
Telethon==1.29.2
typing_extensions==4.7.1
Unidecode==1.3.7
urllib3==2.0.4
vbml==1.1.post1
virtualenv==20.24.5
virtualenvwrapper-win==1.2.7
vk-api==11.9.9
watchfiles==0.18.1
whichcraft==0.6.1
win32-setctime==1.1.0
yarl==1.9.2
zipp==3.16.2
zipstream==1.1.4
zope.event==5.0
zope.interface==6.0
